common: add unit test jobs under gcc
This is intended to match unittest-main-docker-autotest.
Originaly this job was intended to run inside a container
to make better use of resources than VMs. For jobs that
do not run functional tests with SPDK.

Meanwhile container images are not ready for spdk-ci,
this is still executed in VM.

Job was renamed to better match the intent:
- 'main' to 'gcc', other jobs will be similar and have 'clang'
and 'vagrant'
- 'docker' to 'container'

As for the steps, they do not use autorun.sh in order to
skip parts that are required for functional tests - ex. setup.sh.
On old CI setup it was demonstrably faster in execution,
than just full autorun.sh. Yet it might not hold up today.

To keep changes on SPDK repo to minimum, container jobs
will still call functions from tests scripts directly.
At some point it will need to re-evaluted if this is best approach.

One note here is that $rootdir is usually set by autorun.sh
script. Since we are not calling it, we need to set it manually.

OCF is turned on, to match prior CI setup.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>