common: modify nvme job to nvme jobs from Jenkins CI
This expands the previous job config to match more what
was on the jobs from Jenkins CI. Just lumped together to
single job. Moving out ones not related to NVMe.

Enabled ones that just worked for now and the rest
was disabled.

Further work on this can be done in this job, or moved
out to other job/workflow/runner.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>