summary: add attempt number to the url
When GitHub triggers a workflow it gets a specific run_id,
which is unique for the repository. Most runs require just
single attempt, but a failed one can be retriggered manualy
or later in the series via comment on Gerrit.
This results in creating new attempt that uses the same run_id.

Accessing url without specifying a run attempt, just points
to the latest. In order to avoid confusion, be specific.
This avoids old links for failed builds, pointing to
a passing build after false positive comment.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>