gerrit: find relevant GH issue and comment with failed build
Finds the GH issue referred to by the comment, and only checks
if it is still open. Once all the nessecary verification is done
posts a link to failed build and reruns the workflow.

NOTE: ${{ github.token }} that triggers the webhook is limited
in scope to current (spdk-ci) repository, so it cannot post
comments into the spdk proper. Despite the PAT already having
the relevant permissions.
To remedy that a new GH_ISSUES_PAT secret has to be added
with PAT limited only to read/write into spdk proper.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>