gerrit: post comment on false positive
Whenever workflow began to rerun post a message and remove
relevant vote.

Most of the times this workflow is triggered the comments
are irrelevant so no message is needed. Exception might be
when correct message is posted, but to closed or non-existent
GH issue. For this case inform the user.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>