gerrit: fix client_payload mapping
repository_dispatch payload is an object,
it needs to be converted to string before assignment.

Otherwise following error was ocurring:
The template is not valid. .github/workflows/gerrit-false-positives-handler.yml (Line: 16, Col: 19): A mapping was not expecteda

Besides that the "||" statements do not behave as one might expect.
https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/evaluate-expressions-in-workflows-and-actions#example

False positive job won't be able to do anything with empty
client_payload, so throw it out.

Meanwhile per-patch was intended to test latest SPDK
when client_payload is not passed. Until this is figured
out just make that default behaviour and disable passing
client_payload as input for manual runs.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>